rube.core/requirements.txt
vault
selenium
nose
pyzmq<13.0
nose-timer
yanc
pyvirtualdisplay
nose-testconfig
